requirements.txt
"""
Basic list of required modules

    author: Simon Klingler
    date: 3.6.2022
    version: 0.0.1
    license: free
"""
name==version

tkinter==newest
pillow==9.1.0
sqlite3==newest
requests==2.27.1
json==newest
